+++ src/vk/mod.rs
@@ -127,7 +127,7 @@ impl vkPhysicalDevice {
 
         PhysicalDevice {
             handle: self,
-            v_table,
+            v_table: Some(v_table),
         }
     }
 }
@@ -192,4 +192,13 @@ impl PhysicalDeviceMemoryProperties {
 impl ColorComponentFlags {
     /// Contraction of [`R`][Self::R] | [`G`][Self::G] | [`B`][Self::B] | [`A`][Self::A]
     pub const RGBA: Self = Self(Self::R.0 | Self::G.0 | Self::B.0 | Self::A.0);
+}
+
+impl From<Extent2D> for Rect2D {
+    fn from(value: Extent2D) -> Self {
+        Self {
+            offset: Offset2D { x: 0, y: 0 },
+            extent: value,
+        }
+    }
 }